little ui and azure db service prob resolved

+++ app/src/main/java/com/hack36/Interfaces/LocationDao.java
@@ -15,5 +15,5 @@ public interface LocationDao {
     void insert(Location location);
 
     @Query("SELECT * FROM location WHERE time > :time")
-    List<Location> getAllAfter(long time);
+    List<Location> getAllAfter(int time);
 }
